Separated into 3 files
I separated the code into 3 files and fixed some minor code.

Also removed the extra file from the repo to reduce confusion
